+++ ZoranEngine/Core/2D/Camera2D.h
@@ -13,16 +13,16 @@ private:
 	SceneObject2D * sceneObject;
 
 public:
-	Camera2D(std::string name,Vec2D position,Vec2D scale,double rotation);
+	Camera2D(std::string name,Vec2D position,Vec2D scale,float rotation);
 	virtual ~Camera2D();
 
 	virtual MatrixF GetModel()override;
 
-	virtual void Translate(double deltax, double deltay, double deltaz)override;
+	virtual void Translate(float deltax, float deltay, float deltaz)override;
 	virtual void Translate(Vec2D delta)override;
 	virtual void Translate(Vec3D delta)override;
-	virtual void Rotate(double rotx, double roty, double rotz)override;
-	virtual void Rotate(double rotation) override;
+	virtual void Rotate(float rotx, float roty, float rotz)override;
+	virtual void Rotate(float rotation) override;
 	virtual void Rotate(Vec3D euler) override;
 	virtual void Rotate(Quaternion quat)override;
 };
