+++ ZoranEngine/Core/Events/EventBase.h
@@ -9,6 +9,6 @@ public:
 	virtual ~EventBase();
 
 	virtual bool DispatchEvent() = 0;
-	virtual void UpdateEvent(double deltaTime) = 0;
+	virtual void UpdateEvent(float deltaTime) = 0;
 	virtual bool ShouldDispatchEvent() = 0;
 };
